backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
email-validator
python-jose[cryptography]
cryptography
httpx
beautifulsoup4
feedparser
pillow
jinja2
python-multipart
google-generativeai
authlib
itsdangerous
